corrections

+++ EvaGest/Services/Indicators.cs
@@ -6,17 +6,19 @@ namespace EvaGest.Services;
 /// </summary>
 public static class Indicators
 {
-    /// <summary>Share of total takings billed by one worker. Null when the period had no sales.</summary>
-    public static decimal? WorkPercentage(int workerCents, int periodTotalCents)
+    /// <summary>Share of total takings billed by one worker. Null when the period had no sales.
+    /// Takes long because period totals are summed as long: clamping them down to int at
+    /// the call site used to silently report 100 % instead of overflowing.</summary>
+    public static decimal? WorkPercentage(long workerCents, long periodTotalCents)
         => periodTotalCents == 0 ? null : workerCents * 100m / periodTotalCents;
 
     /// <summary>Share of a worker's takings that came from products.
     /// Null when that worker had no sales in the period.</summary>
-    public static decimal? ProductsPercentage(int productsCents, int workerCents)
+    public static decimal? ProductsPercentage(long productsCents, long workerCents)
         => workerCents == 0 ? null : productsCents * 100m / workerCents;
 
     /// <summary>Average spend per completed visit. Null when the client has no completed visits.</summary>
-    public static decimal? AveragePerVisit(int totalSpentCents, int completedVisits)
+    public static decimal? AveragePerVisit(long totalSpentCents, int completedVisits)
         => completedVisits == 0 ? null : totalSpentCents / 100m / completedVisits;
 
     /// <summary>
